Some things didn't quite merge as cleanly as they could have, went ahead and patched em up.

+++ Assets/Scripts/Modes/Tactical/TacticalUtilities.cs
@@ -626,7 +626,7 @@ static class TacticalUtilities
         return;
     }
 
-    static public void MutuallyDamageUnits(Actor_Unit source, int damage)
+    static public void MutuallyDamageUnits(Actor_Unit source, double damage)
     {
         if (linkedUnits == null)
             return;
